pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "privenergybench"
version = "0.1.0"
description = "Utility-privacy-energy benchmarking toolkit for PyTorch classifiers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "PyYAML>=6.0",
  "scikit-learn>=1.4",
  "torch>=2.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
